Dribble stats (#111)
* Fixed formatting of check_rattletrap_version.py, updated rattletrap versions.

* rename basic stats, fix typo.

* Fixed boosts pandas false positive warning.

* Added saltie_hit save.

* Added possession analysis.

* remove old rattle trap

* ignore iml

* added new test replays

* adding more tests

* added dribble replays for test

* added dribble data to proto

* adjusted stats

* added event creator

* added dribble test

* moved events into new directory

* Added documentation

* added inits to the directories

This should allow the files to be found when running

* fixed broken tests

* fix imports for event_creator

* added dribble detection

* fix tests so they pass

* Added all of the carry stats

Added carry stats

* finished all dribble stats

* fix error with testing

* some field renaming

* fixed division by zero

* added dribble tests

(need to figure out why there is a nan)

* Ignored files that do not matter

This should hopefully also increase our coverage % by a ton

* updated ignored files again

* fixed ignored paths

* ignore all utils

* Added the merging of dribbles.

* rename basic stats, fix typo.

* Fixed boosts pandas false positive warning.

* Added saltie_hit save.

* Added possession analysis.

* Added per-possession stats.

* Fixed "might not be assigned" warning.

* Added proto to VC

* fixed bug in fix.

* pass the frickin tests.

* added a comment

* Moved stuff to player and team stats.

* Added saves test.

* Added todo.

* Remove 0 saves from test.

* progress on carry enhancements

* Added distance along path

* Cut off dribbles early if they are hit by a player.

* fixed some documentation and naming conventions

* modifications to carries for cutting them off

* added a new test

Fixed some issues with merging

* added more changes to dribbles

Fixed the missing dribble.
Now I have one extra I need to remove

* fixed and passes all unit tests!

* fixed issue with merge

* fix import error

* fix error cases

zero dribbles
people leaving.

* added a test for zero carries

* Finished merge

* moved kickoff stats to being in events folder

* Added files for local replays

* fixed some imports

* requireing a minimum pandas

* Fixed merging of ball carries

Carries are now correctly merged and split

* Fixed modified answers

* oopsiedoopsie

* fixed crash issue

Fixes issue where the carry frame exists but players dont have any dribbles.
Fix issue during dribble merges

* Update cloud_parser

* addressed increased dribble time

:(
changing the merge increased dribble time
