+++ api/stats/events.proto
@@ -1,6 +1,7 @@
 syntax = "proto2";
 
 import "api/player_id.proto";
+import "api/stats/stats.proto";
 import "api/stats/extra_mode_stats.proto";
 
 package api.stats;
@@ -81,6 +82,19 @@ message FiftyFifty {
 message Kickoff {
     optional int32 start_frame_number = 1;
     optional int32 end_frame_number = 2;
+}
+
+// Events where the car carried the ball on its top.
+message BallCarry {
+    optional int32 start_frame_number = 1;
+    optional int32 end_frame_number = 2;
+    optional api.PlayerId player_id = 3;
+    optional bool has_flick = 4; // True if the dribble ended in a flick
+    optional float carry_time = 5;
+    optional float straight_line_distance = 6; // This is just the start/end distance traveled. It is NOT distance along path
+
+    // Stats
+    optional DetailedCarryStats carry_stats = 7;
 }
 
 message RumbleItemEvent {
