+++ api/stats/game_stats.proto
@@ -14,4 +14,5 @@ message GameStats {
     repeated api.stats.Kickoff kickoffs = 5;
     repeated api.stats.RumbleItemEvent rumble_items = 6;
     repeated api.stats.KickoffStats kickoff_stats = 7;
+    repeated api.stats.BallCarry ball_carries = 8;
 }
