+++ api/stats/player_stats.proto
@@ -65,10 +65,25 @@ message PlayerStats {
     optional RelativePositioning relative_positioning = 10;
     optional api.stats.PerPossessionStats per_possession_stats = 11;
     optional api.stats.RumbleStats rumble_stats = 12;
+    optional CarryDribbles ball_carries = 13;
 }
 
 message Controller {
     optional bool is_keyboard = 1;
     optional float analogue_steering_input_percent = 2;
     optional float analogue_throttle_input_percent = 3;
+}
+
+// Stats for carrying
+message CarryDribbles {
+    optional int32 total_carries = 1;
+    optional int32 total_flicks = 2;
+    optional float longest_carry = 3;
+    optional float furthest_carry = 4;
+    optional float total_carry_time = 5;
+    optional float average_carry_time = 6;
+    optional float fastest_carry_speed = 7; // Your speed along the ground while dribbling the ball
+    optional float total_carry_distance = 8; // Just adding up the start/end distances.  It is NOT distance along path
+
+    optional DetailedCarryStats carry_stats = 9;
 }
