+++ codecov.yml
@@ -31,3 +31,9 @@ comment:
   layout: "header, diff"
   behavior: default
   require_changes: no
+
+ingore:
+  - ".*/tests/**/*"
+  - "setup.py"
+  - "init.py"
+  - "utils/.*"
